refactor: server into files routing/handlers/middleware
